requirements.txt
django>=5.0
djangorestframework>=3.15
gunicorn>=21.2
pdfplumber
spacy
